Happiness meter has been added to the project. The happiness meter decreases by 10% every hour for every single user. The happiness meter can be increased by buying food for your animal.
Currently this is not balanced at all, feel free to fiddle with the prices/rate of decrease of happiness.

The happiness cannot go below 0 or above 100, and buying food for your animal that would put you past 100 is currently wasted (you lose your money but happiness does not go over 100).